chore: format thymeleaf templates with prettier
Adds a command in `e2e` to use `prettier` as a formatter for Thymeleaf templates: `yarn format:backend-templates`.
This is not enforced by a build check, it's just as an help for developers.

One silver-lining of such a formatter is that it can catch HTML errors.
For example, there was a missing closing `div` in the existing `home` template, and it throwed:

```
[error] ../backend/src/main/resources/templates/home.html: SyntaxError: Unexpected closing tag "main". It may happen when the tag has already been closed by another tag. For more info see https://www.w3.org/TR/html5/syntax.html#closing-elements-that-have-implied-end-tags (44:3)
```

I think this could be helpful in the future, when the templates will get more complex.