feat: integrate map in home
Note that the JS has been modified, and that a slug has been added in the HTML, too.
Also note that there are some parts that are missing in the territory model, and that some parts of the HTML are thus commented out or replaced by fake data (see the TODO).